TST: Move RichSignal to a common location
