+++ tests/test_plugins.py
@@ -5,19 +5,13 @@ from pydm.PyQt.QtGui import QWidget
 from typhon.plugins.core import (SignalPlugin, SignalConnection,
                                  register_signal)
 
+from .conftest import RichSignal
+
+
 class WritableWidget(QWidget, PyDMWritableWidget):
     """Simple Testing Widget"""
     pass
 
-
-class RichSignal(Signal):
-
-    def describe(self):
-        return {self.name : {'enum_strs': ('a', 'b', 'c'),
-                             'precision': 2,
-                             'units': 'urad',
-                             'dtype': 'number',
-                             'shape': []}}
 
 class DeadSignal(Signal):
     subscribable = False
